+++ rewrite/src/platform/freestand/getstrlen.h
@@ -1,7 +1,7 @@
 #pragma once
 
 size_t getstrlen(char *s) {
-	size_t n = 0;
+	unsigned long n = 0;
 
 	while (s[n] != '\0')
 		n++;
