+++ rewrite/src/state/variables.h
@@ -4,8 +4,8 @@
 
 typedef SHashMap VarMap;
 
-static size_t hashVar(VarMap *m, char *str)              { return mapHash(m, str); }
-static VarMap initVars(size_t cap)                       { return initHashMap(cap); }
-static void   putVar(VarMap *m, char *key, void *value)  { hashMapPut(m, key, value); }
-static void  *getVar(VarMap *m, char *key)               { return hashMapGet(m, key); }
-static void   freeVars(VarMap *m)                        { hashMapFreeMap(m); }
+static unsigned long hashVar(VarMap *m, char *str)		{ return mapHash(m, str); }
+static VarMap initVars(unsigned long cap)			{ return initHashMap(cap); }
+static void   putVar(VarMap *m, char *key, void *value)		{ hashMapPut(m, key, value); }
+static void  *getVar(VarMap *m, char *key)			{ return hashMapGet(m, key); }
+static void   freeVars(VarMap *m)				{ hashMapFreeMap(m); }
